Split the driver into controller + node binaries
The node and controller parts of the driver are fundamentally different.
The node part runs in a DaemonSet and can *only* run on a Hetzner Cloud
server instance. The controller by default runs in a Deployment, but it
can run anywhere, even outside of the cluster in advanced scenarios.

Thus the environmental requirements and assumptions for these two
components also differ. By splitting into two separate binaries, it's
easier to reason with these differences, and properly specialize them
appropriately.

This initial work does the split, and carves out an "app" package that
has the common initialization code (clients, loggers, metrics, etc). No
behaviour changes have been introduced yet, the node and the controller
should function exactly as they did before - just in different binaries.

Follow-up work will take advantage of this split.
