+++ deploy/kubernetes/controller/deployment.yaml
@@ -36,6 +36,7 @@ spec:
       - name: hcloud-csi-driver
         image: hetznercloud/hcloud-csi-driver:latest
         imagePullPolicy: Always
+        command: [/bin/hcloud-csi-driver-controller]
         env:
         - name: CSI_ENDPOINT
           value: unix:///run/csi/socket
